requirements.txt
notion-client
pyodbc
python-slugify
